+++ apps/src/componentLibrary/dropdown/simpleDropdown/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.2] ()
+
+* added `errorMessage` prop to `SimpleDropdown`
+* added error state border color to `SimpleDropdown` when `errorMessage` exists
+* added `helperMessage` and `helperIcon` props to `SimpleDropdown`
+
 ## [0.6.1] ()
 
 * improved `SimpleDropdown` down arrow icon RTL support